Move Handle struct to new module
